Proxy returns formatted response or redirection.
We want to have the proxy return better errors for humans in some cases.
Particularly, in the case of Probo SASS, we want the probo frontend to
be redirected to a nice error page. If however there is no front end and
the user is visiting the proxy URL via a browser, they should also see a
nice(r) error page. To handle this, configuration settings were added
that allow the proxy to be configured with a redirection URL for error
handling and a string of custom 404 error HTML. We inspect the request's
headers to determine if the response should be HTML (redirection or
HTML for a browser) or JSON (cli) or a plaintext response (no accept
headers sent).